✨ Move validation middleware into WorkflowComposer with UseWorkflowComposer()
