+++ src/CompositionGateway/Program.cs
@@ -1,4 +1,3 @@
-using CompositionGateway;
 using ITOps.Shared.EndpointConfiguration;
 using ITOps.Shared.Sqlite;
 using ServiceComposer.AspNetCore;
@@ -49,7 +48,7 @@ if (app.Environment.IsDevelopment())
     app.MapOpenApi();
 }
 
-app.UseMiddleware<WorkflowValidationExceptionMiddleware>();
+app.UseWorkflowComposer();
 
 app.UseCors();
