+++ src/WorkflowComposer/WorkflowComposer.csproj
@@ -9,9 +9,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="10.0.8" />
-      <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="10.0.8" />
-      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.8" />
+      <FrameworkReference Include="Microsoft.AspNetCore.App" />
     </ItemGroup>
 
 </Project>
